feat: remove transaction manager
