+++ libs/deal-ticket/src/components/deal-ticket/deal-ticket-status.tsx
@@ -1,31 +1,22 @@
 import { useCallback, useMemo } from 'react';
 import type { ReactNode } from 'react';
-import type { VegaTxState, OrderSubmissionBody } from '@vegaprotocol/wallet';
+import type { VegaTxState } from '@vegaprotocol/wallet';
 import {
-  VegaTxStatus,
   WalletError,
   useVegaWallet,
   useVegaWalletDialogStore,
   ClientErrors,
-  useVegaTransactionStore,
 } from '@vegaprotocol/wallet';
-import { DealTicket } from './deal-ticket';
-import type { MarketDealTicket } from '@vegaprotocol/market-list';
-import { useOrderSubmit, OrderFeedback } from '@vegaprotocol/orders';
 import * as Schema from '@vegaprotocol/types';
 import { Button, Icon, Intent } from '@vegaprotocol/ui-toolkit';
 import { t } from '@vegaprotocol/react-helpers';
-
-export interface DealTicketManagerProps {
-  market: MarketDealTicket;
-  children?: ReactNode | ReactNode[];
-}
 
 interface ErrorContentProps {
   transaction: VegaTxState;
   reset: () => void;
 }
-const ErrorContent = ({ transaction, reset }: ErrorContentProps) => {
+
+export const ErrorContent = ({ transaction, reset }: ErrorContentProps) => {
   const { openVegaWalletDialog } = useVegaWalletDialogStore((store) => ({
     openVegaWalletDialog: store.openVegaWalletDialog,
   }));
@@ -61,23 +52,6 @@ const ErrorContent = ({ transaction, reset }: ErrorContentProps) => {
     }
     return null;
   }, [transaction, reconnect]);
-};
-
-export const DealTicketManager = ({
-  market,
-  children,
-}: DealTicketManagerProps) => {
-  const create = useVegaTransactionStore((state) => state.create);
-  return (
-    <>
-      {children || (
-        <DealTicket
-          market={market}
-          submit={(orderSubmission) => create({ orderSubmission })}
-        />
-      )}
-    </>
-  );
 };
 
 export const getOrderDialogTitle = (
