+++ libs/deal-ticket/src/components/deal-ticket/index.ts
@@ -1,7 +1,7 @@
 export * from './deal-ticket-amount';
 export * from './deal-ticket-container';
 export * from './deal-ticket-limit-amount';
-export * from './deal-ticket-manager';
+export * from './deal-ticket-status';
 export * from './deal-ticket-market-amount';
 export * from './deal-ticket';
 export * from './expiry-selector';
